requirements.txt
south
python-openid
django_openid_auth
threadedcomments
